Update the old SSMS for Oracle common errors and how to fix it
Update the old SSMS for Oracle common errors and how to fix it.
I have removed the old document "Update the old SSMS for Oracle common errors and how to fix it" and replaced it with a new document SSMS issues and workarounds when migrating Oracle database. It contains more error codes, explanations and possible workarounds than the original document